refactor: moved all dto's to its own folders
