+++ src/main/java/com/ssdev/mypet/domain/auth/dto/AuthLoginDto.java
@@ -1,3 +1,3 @@
-package com.ssdev.mypet.domain.auth;
+package com.ssdev.mypet.domain.auth.dto;
 
 public record AuthLoginDto (String username, String password) {}
