+++ src/main/java/com/ssdev/mypet/domain/auth/dto/AuthRegisterDto.java
@@ -1,3 +1,3 @@
-package com.ssdev.mypet.domain.auth;
+package com.ssdev.mypet.domain.auth.dto;
 
 public record AuthRegisterDto(String username) {}
